move the only other function from solver to utils
