+++ lib/solver.mjs
@@ -1,29 +1,4 @@
-import { getPossibilities, getFirstEmptyCell, boardError, setCharAt, memoize } from './utils'
-
-const memoizedGetGuesses = memoize(getGuesses)
-const memoizedSetCharAt = memoize(setCharAt)
-
-function getGuesses(board) {
-  const guesses = []
-  for (let r = getFirstEmptyCell(board); r !== -1; r = getFirstEmptyCell(board, r + 1)) {
-    const possibilities = getPossibilities(board, r)
-    if (!possibilities) {
-      // At least one empty cell has no possibilities so the board doesn't worth exploring
-      return
-    }
-    if (possibilities.length === 1) {
-      // One cell with a single possiblity is detected so just use that one
-      return [{ r, possibilities }]
-    }
-    guesses.push({ r, possibilities })
-  }
-
-  if (guesses.length >= 2) {
-    // sort the guesses so the ones with less possibilities end up last so they'll be at the end of unsolvedBoards
-    return guesses.sort((a, b) => b.possibilities.length - a.possibilities.length)
-  }
-  return guesses
-}
+import { memoizedGetGuesses, getFirstEmptyCell, boardError, memoizedSetCharAt } from './utils'
 
 export function solve(input, maxAnswers = 2) {
   if (maxAnswers <= 0) {
