+++ lib/utils.mjs
@@ -123,6 +123,31 @@ export function getPossibilities(board, r) {
   return vals.filter(Boolean)
 }
 
+export const memoizedGetGuesses = memoize(getGuesses)
+export const memoizedSetCharAt = memoize(setCharAt)
+
+function getGuesses(board) {
+  const guesses = []
+  for (let r = getFirstEmptyCell(board); r !== -1; r = getFirstEmptyCell(board, r + 1)) {
+    const possibilities = getPossibilities(board, r)
+    if (!possibilities) {
+      // At least one empty cell has no possibilities so the board doesn't worth exploring
+      return
+    }
+    if (possibilities.length === 1) {
+      // One cell with a single possiblity is detected so just use that one
+      return [{ r, possibilities }]
+    }
+    guesses.push({ r, possibilities })
+  }
+
+  if (guesses.length >= 2) {
+    // sort the guesses so the ones with less possibilities end up last so they'll be at the end of unsolvedBoards
+    return guesses.sort((a, b) => b.possibilities.length - a.possibilities.length)
+  }
+  return guesses
+}
+
 export function getFirstEmptyCell(board, from = 0) {
   return board.indexOf(' ', from)
 }
